move parse hgnc func to util
